solve tests
