+++ tests/metrics/test_converters.py
@@ -6,7 +6,7 @@ import torch.distributed as dist
 import tests.base.utils as tutils
 from pytorch_lightning.metrics.converters import _apply_to_inputs, _apply_to_outputs, \
     _convert_to_tensor, _convert_to_numpy, _numpy_metric_conversion, \
-    _tensor_metric_conversion, _sync_ddp, tensor_metric, numpy_metric
+    _tensor_metric_conversion, _sync_ddp_if_available, tensor_metric, numpy_metric
 
 
 def test_apply_to_inputs():
@@ -106,7 +106,7 @@ def test_sync_reduce_ddp():
 
     tensor = torch.tensor([1.], device='cuda:0')
 
-    reduced_tensor = _sync_ddp(tensor)
+    reduced_tensor = _sync_ddp_if_available(tensor)
 
     assert reduced_tensor.item() == dist.get_world_size(), \
         'Sync-Reduce does not work properly with DDP and Tensors'
@@ -118,7 +118,7 @@ def test_sync_reduce_simple():
     """Make sure sync-reduce works without DDP"""
     tensor = torch.tensor([1.], device='cpu')
 
-    reduced_tensor = _sync_ddp(tensor)
+    reduced_tensor = _sync_ddp_if_available(tensor)
 
     assert torch.allclose(tensor,
                           reduced_tensor), 'Sync-Reduce does not work properly without DDP and Tensors'
